uefi-raw: Move common USB structures
Move USB structures used in multiple protocol interfaces to the general usb
module.
